UI Changes.

+++ app/src/main/java/com/smartsettings/ai/resources/db/dbModels.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "smart_settings")
 data class SmartSettingDBModel(
 
-    @PrimaryKey(autoGenerate = true) var id: Int? = null,
+    @PrimaryKey(autoGenerate = true) var id: Long? = null,
 
     @ColumnInfo(name = "name") var name: String,
